refactor on images

+++ src/api/card/views/cards.py
@@ -7,7 +7,11 @@ from django.views.decorators.csrf import csrf_exempt
 from drf_spectacular.utils import extend_schema
 from api.card import docs
 from api.card.models import Card
-from api.card.serializers.card import CardSerializer
+from api.card.serializers.card import (
+    CardSerializer,
+    CardSerializerList,
+    CardIllustrationSerializer,
+)
 
 
 @extend_schema(**docs.card_list)
@@ -15,7 +19,7 @@ from api.card.serializers.card import CardSerializer
 @api_view(("GET",))
 @permission_classes((AllowAny,))
 def card_list(request: Request) -> Response:
-    card_list_response = CardSerializer(Card.objects.all(), many=True)
+    card_list_response = CardSerializerList(Card.objects.all(), many=True)
     return Response(card_list_response.data, status=status.HTTP_200_OK)
 
 
@@ -24,9 +28,10 @@ def card_list(request: Request) -> Response:
 @api_view(("GET",))
 @permission_classes((AllowAny,))
 def card_detail(request: Request, card_id: int) -> Response:
-    card = Card.objects.get(id=card_id)
-    card_response = CardSerializer(card)
-    return Response(card_response.data, status=status.HTTP_200_OK)
+    if card := Card.objects.filter(id=card_id).first():
+        card_response = CardSerializer(card)
+        return Response(card_response.data, status=status.HTTP_200_OK)
+    return Response("Card not found", status=status.HTTP_404_NOT_FOUND)
 
 
 @extend_schema(**docs.create_card)
@@ -46,12 +51,13 @@ def create_card(request: Request) -> Response:
 @api_view(("PUT",))
 @permission_classes((IsAuthenticated,))
 def update_card(request: Request, card_id: int) -> Response:
-    card = Card.objects.get(id=card_id)
-    card_serializer = CardSerializer(card, data=request.data)
-    if card_serializer.is_valid():
-        card_serializer.save()
-        return Response(card_serializer.data, status=status.HTTP_200_OK)
-    return Response(card_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    if card := Card.objects.filter(id=card_id).first():
+        card_serializer = CardSerializer(card, data=request.data)
+        if card_serializer.is_valid():
+            card_serializer.save()
+            return Response(card_serializer.data, status=status.HTTP_200_OK)
+        return Response(card_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    return Response("Card not found", status=status.HTTP_404_NOT_FOUND)
 
 
 @extend_schema(**docs.delete_card)
@@ -59,6 +65,54 @@ def update_card(request: Request, card_id: int) -> Response:
 @api_view(("DELETE",))
 @permission_classes((IsAdminUser,))
 def delete_card(request: Request, card_id: int) -> Response:
-    card = Card.objects.get(id=card_id)
-    card.delete()
-    return Response(status=status.HTTP_204_NO_CONTENT)
+    if card := Card.objects.filter(id=card_id).exists():
+        card.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
+    return Response("Card not found", status=status.HTTP_404_NOT_FOUND)
+
+
+@extend_schema(**docs.illustration_list)
+@csrf_exempt
+@api_view(("GET",))
+@permission_classes((AllowAny,))
+def card_illustrations(request: Request, card_id: int) -> Response:
+    if card := Card.objects.filter(id=card_id).first():
+        serializer = CardIllustrationSerializer(card.illustrations, many=True)
+        return Response(serializer.data, status=status.HTTP_200_OK)
+    return Response("Card not found", status=status.HTTP_404_NOT_FOUND)
+
+
+@extend_schema(**docs.create_card_illustration)
+@csrf_exempt
+@api_view(("POST",))
+@permission_classes((IsAuthenticated,))
+def create_card_illustration(request: Request, card_id: int) -> Response:
+    serializer = CardSerializer(data=request.data)
+    if serializer.is_valid():
+        serializer.save()
+        return Response(serializer.data, status=status.HTTP_201_CREATED)
+    return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+
+@extend_schema(**docs.card_illustration_detail)
+@csrf_exempt
+@api_view(("GET",))
+@permission_classes((AllowAny,))
+def card_illustration_detail(
+    request: Request, card_id: int, illustration_id: int
+) -> Response:
+    if card := Card.objects.filter(id=card_id).exists():
+        if illustration := card.illustrations.filter(id=illustration_id).exists():
+            serializer = CardIllustrationSerializer(illustration, many=True)
+            return Response(serializer.data, status=status.HTTP_200_OK)
+        return Response("Illustration not found", status=status.HTTP_404_NOT_FOUND)
+    return Response("Card not found", status=status.HTTP_404_NOT_FOUND)
+
+
+@extend_schema(**docs.illustration_list)
+@csrf_exempt
+@api_view(("GET",))
+@permission_classes((AllowAny,))
+def illustration_list(request: Request) -> Response:
+    serializer = CardIllustrationSerializer(Card.objects.all(), many=True)
+    return Response(serializer.data, status=status.HTTP_200_OK)
